refactor: refine codes; fix errors

+++ Data Structures/Persistent Segment Tree/Kth Number Hackerrank/Solution.cpp
@@ -53,7 +53,7 @@ Node *merge(Node *leftNode, Node *rightNode, bool in_place = 0, Node *Current = 
 	}
 }
 
-Node *insert(Node *node, int left, int right, int i, ll val) // Update for a single version
+Node *insert(Node *node, ll left, ll right, ll i, ll val) // Update for a single version
 {
 	if (left <= i && right >= i)
 	{
@@ -71,7 +71,7 @@ Node *insert(Node *node, int left, int right, int i, ll val) // Update for a sin
 	return node;
 }
 
-ll query(Node *Current, int left, int right, int k)
+ll query(Node *Current, ll left, ll right, ll k)
 {
 	if (left == right)
 		return left;
@@ -85,12 +85,12 @@ ll query(Node *Current, int left, int right, int k)
 
 // Interface
 
-Node *insert(Node *Current, int i, ll val)
+Node *insert(Node *Current, ll i, ll val)
 {
 	return insert(Current, 0, N, i, val);
 }
 
-ll query(Node *Current, int k)
+ll query(Node *Current, ll k)
 {
 	ll ans = query(Current, 0, N, k);
 	return ans;
