+++ Data Structures/Persistent Segment Tree/Persistent Bookcase/Solution.cpp
@@ -71,7 +71,7 @@ Node *merge(Node *leftNode, Node *rightNode, bool in_place = 0, Node *Current = 
 	}
 }
 
-Node *insert(Node *node, int left, int right, int i, int j, int state) // Update for a single version
+Node *insert(Node *node, ll left, ll right, ll i, ll j, int state) // Update for a single version
 {
 	if (left <= i && right >= i)
 	{
@@ -106,7 +106,7 @@ Node *insert(Node *node, int left, int right, int i, int j, int state) // Update
 	return node;
 }
 
-ll query(Node *Current, int left, int right, int leftQuery, int rightQuery)
+ll query(Node *Current, ll left, ll right, ll leftQuery, ll rightQuery)
 {
 	// [left, right] doesn't intersect with [leftQuery, rightQuery]
 	if (Current == nullptr || left > rightQuery || right < leftQuery)
@@ -123,12 +123,12 @@ ll query(Node *Current, int left, int right, int leftQuery, int rightQuery)
 
 // Interface
 
-Node *insert(Node *Current, int i, int j, int state)
+Node *insert(Node *Current, ll i, ll j, int state)
 {
 	return insert(Current, 1, N, i, j, state);
 }
 
-ll query(Node *Current, int left, int right)
+ll query(Node *Current, ll left, ll right)
 {
 	ll ans = query(Current, 1, N, left, right);
 	return ans;
